+++ dev/test.js
@@ -35,11 +35,13 @@ $('.click-me-1').on('click',function(e){
 });
 
 var popbox2 = new Popbox({
-    max_width:800,
-    content:'<p>test masses of content</p><img src="assets/img/test-image-2.jpg" />',
-    title:'test 2',
+    add_class: 'popbox-phat-fit',
+    //max_width:800,
+    content:'<img src="assets/img/test-image-2.jpg" width="653" height="803" /><div style="font-size:12px; padding: 16px 0 0 0;"><p>test masses of content with more content than you might need masses of content with more content than you might need masses of content with more content than you might need masses of content with more content than you might need</p><p>Next content</p></div>',
+    //title:'test 2',
     animation:'zoom',
-    max_height:true,
+    //max_height:true,
+    fit:true,
     loading_text:'<img src="assets/img/loading.gif" width="32" height="32" alt="" />'
 });
 
@@ -51,7 +53,7 @@ $('.click-me-2').on('click',function(e){
 var popbox3 = new Popbox({
     max_width:700,
     //max_height:true,
-    aspect_fit:true,
+    fit:true,
     content:'<p>test more content</p>',
     animation:'zoom',
     cache:true,
@@ -179,4 +181,6 @@ var popbox_breakage = new Popbox({
     }
 });
 
-popbox_breakage.open();
+setTimeout(function(){
+    popbox_breakage.open();
+},500);